WEEK_9/vaultguard-api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vaultguard-api"
version = "0.1.0"
description = "AI-powered maintenance intelligence platform for VaultGuard smart safe billing validation"
authors = [
    {name = "Pedro Biai", email = "pedro@example.com"}
]
dependencies = [
    "mcp>=1.0.0",
    "fastmcp>=0.9.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "openai>=1.0.0",
    "langsmith>=0.1.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.1.0",
    "google-api-python-client>=2.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.25.0"
]
requires-python = ">=3.9"

[project.scripts]
vaultguard-server = "src.server:main"
vaultguard-api = "src.api_server:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]
